refactor: reworked app, brugerConsrtuctor, modal, burgerIngredients
