+++ src/components/BurgerIngredient/BurgerIngredient.module.css
@@ -1,0 +1,21 @@
+.ingredient {
+	width: 272px;
+	height: 208px;
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	position: relative;
+}
+.ingredient__image {
+	width: 240px;
+	height: 120px;
+	object-fit: cover;
+	object-position: center;
+}
+.ingredient__price {
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	gap: 8px;
+	cursor: pointer;
+}
