+++ src/components/ChosenElement/ChosenElement.module.css
@@ -1,0 +1,10 @@
+.filling {
+	display: grid;
+	grid-template-rows: repeat(auto, 1fr);
+	grid-template-columns: 24px 1fr;
+	column-gap: 8px;
+	align-items: center;
+	overflow-y: scroll;
+	max-height: 464px;
+	row-gap: 16px;
+}
