--- src/components/Modal/Modal.module.css
@@ -1,10 +1,0 @@
-.popup {
-	position: absolute;
-	display: flex;
-	flex-direction: column;
-	justify-content: center;
-	align-items: center;
-	width: 720px;
-	background-color: var(--background);
-	border-radius: 40px;
-}
