+++ src/components/ModalOverlay/ModalOverlay.module.css
@@ -1,9 +1,19 @@
-.container {
+.popup {
 	position: fixed;
 	display: flex;
 	justify-content: center;
 	align-items: center;
-	top: 50%;
-	left: 50%;
-	z-index: 1;
+	top: 0;
+	width: 100vw;
+	height: 100vh;
+}
+.container {
+	display: flex;
+	position: relative;
+	width: 720px;
+	flex-direction: column;
+	justify-content: center;
+	align-items: center;
+	background-color: var(--background);
+	border-radius: 40px;
 }
