+++ src/components/TotalCost/TotalCost.module.css
@@ -1,0 +1,5 @@
+.price {
+	display: flex;
+	flex-direction: row-reverse;
+	align-items: center;
+}
